+++ src/main/scala/engine/static_eval/StaticEvaluator.scala
@@ -4,7 +4,7 @@ import chess._
 import engine.static_eval.simplified_evaluation_function.Simplified_Evaluation_Function
 
 trait StaticEvaluator:
-  def evaluate(board: Board): Int
+  def evaluate(board: Board): Double
 
 object StaticEvaluator: 
   given StaticEvaluator = Simplified_Evaluation_Function
